requirements.txt
torch
numpy
lightning
stable-baselines3
gymnasium-robotics